requirements.txt
python-telegram-bot>=20.0
yt-dlp>=2023.3.1
logging
os
re
asyncio
shutil
